requirements-production.txt
# Mainza AI Production Requirements
# Context7 MCP Standards Implementation

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and Caching
neo4j==5.14.1
redis==5.0.1
aioredis==2.0.1
psycopg2-binary==2.9.9

# Security and Authentication
cryptography==41.0.8
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Performance and Monitoring
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# HTTP and Networking
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Data Processing and ML
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.1
transformers==4.36.2

# Audio Processing
whisper==1.1.10
TTS==0.22.0
librosa==0.10.1
soundfile==1.0.4
pydub==0.25.1

# LiveKit Integration
livekit==0.10.1
livekit-api==0.5.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
tenacity==8.2.3

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-mermaid2-plugin==1.1.1

# Production Server
gunicorn==21.2.0
supervisor==4.2.5

# Monitoring and Observability
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Environment and Configuration
pydantic-settings==2.1.0
dynaconf==3.2.4

# Async and Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aiocache==0.12.2

# Data Validation and Serialization
marshmallow==3.20.1
cerberus==1.3.5
jsonschema==4.20.0

# Networking and Communication
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Time and Date Handling
python-dateutil==2.8.2
pytz==2023.3.post1
arrow==1.3.0

# File and Path Handling
pathlib2==2.3.7.post1
watchdog==3.0.0

# Encryption and Hashing
hashlib-compat==1.0.1
argon2-cffi==23.1.0

# Memory and Resource Management
memory-profiler==0.61.0
pympler==0.9
objgraph==3.6.0

# System Integration
systemd-python==235
dbus-python==1.3.2

# Error Tracking and Debugging
rollbar==0.16.3
bugsnag==4.6.1

# Performance Profiling
py-spy==0.3.14
line-profiler==4.1.1
cProfile-tools==1.0.0

# Load Testing
locust==2.17.0
artillery==1.7.9

# Container Support
docker==6.1.3
kubernetes==28.1.0

# Cloud Services (Optional)
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Message Queues (Optional)
celery==5.3.4
kombu==5.3.4
redis-py-cluster==2.1.3

# Search and Indexing (Optional)
elasticsearch==8.11.0
whoosh==2.7.4

# Image Processing (Optional)
Pillow==10.1.0
opencv-python==4.8.1.78

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Graph Processing
networkx==3.2.1
igraph==0.11.3

# Configuration Management
configparser==6.0.0
toml==0.10.2
yaml==0.2.5

# Backup and Archiving
tarfile==0.1.0
zipfile36==0.1.3

# Logging and Audit
audit-log==1.0.0
python-audit==1.1.0

# Health Checks
health-check==1.3.3
py-healthcheck==1.10.1

# Rate Limiting
slowapi==0.1.9
limits==3.6.0

# Circuit Breaker
pybreaker==1.0.1
circuit-breaker==1.0.0

# Retry Logic
retrying==1.3.4
backoff==2.2.1

# Caching
cachetools==5.3.2
diskcache==5.6.3

# Serialization
pickle5==0.0.12
dill==0.3.7
joblib==1.3.2

# Compression
zstandard==0.22.0
lz4==4.3.2

# Validation
validators==0.22.0
email-validator==2.1.0.post1

# UUID and ID Generation
uuid==1.30
shortuuid==1.0.11

# Timezone Handling
tzlocal==5.2
zoneinfo==0.2.1

# Process Management
supervisor==4.2.5
daemon==1.2

# System Monitoring
py-cpuinfo==9.0.0
GPUtil==1.4.0

# Development Tools (for production debugging)
ipython==8.17.2
ipdb==0.13.13
pdb-attach==3.2.0

# API Documentation
fastapi-users==12.1.2
fastapi-limiter==0.1.5
fastapi-cache2==0.2.1

# WebSocket Support
python-socketio==5.10.0
websocket-client==1.6.4

# Background Tasks
rq==1.15.1
dramatiq==1.15.0

# Metrics and Analytics
statsd==4.0.1
datadog==0.48.0

# Feature Flags
flagsmith==3.5.0
unleash-client==5.11.1

# A/B Testing
split-io==8.3.0

# Internationalization
babel==2.13.1
gettext==4.0

# Template Engine
jinja2==3.1.2
mako==1.3.0

# XML Processing
lxml==4.9.3
xmltodict==0.13.0

# CSV Processing
csvkit==1.1.1
pandas-profiling==3.6.6

# Excel Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Processing
PyPDF2==3.0.1
reportlab==4.0.7

# QR Code Generation
qrcode==7.4.2
python-barcode==0.15.1

# Color and Styling
colorama==0.4.6
termcolor==2.4.0

# Progress Bars
tqdm==4.66.1
alive-progress==3.1.5

# CLI Tools
click-completion==0.5.2
click-help-colors==0.9.2

# Environment Detection
python-decouple==3.8
environs==10.0.0

# Dependency Injection
dependency-injector==4.41.0
punq==0.6.2

# State Machines
transitions==0.9.0
python-statemachine==2.1.2

# Event Sourcing
eventsourcing==9.2.21

# CQRS Pattern
axon-python==0.1.0

# Domain-Driven Design
ddd-python==0.1.0

# Clean Architecture
clean-architecture==0.1.0

# Hexagonal Architecture
ports-and-adapters==0.1.0